follow 'helper behavior' pattern that Blacklight uses, to support overriding helper methods; tweak obj var name in #facet_field_names_for_advanced_search to match method name
